[UPD] Renamed settings window's JS to TSX
